requirements.txt
Pillow==7.1.2
numpy==1.18.4